MemoryRegion -> ByteRegion; add Merkle proof-related methods to protocol

+++ docs/protocol/merkle_proof.go
@@ -10,13 +10,13 @@ type MerkleProof struct {
 }
 
 // splits the byte sequence into chunks of specific size
-func Split(data []byte, chunkSize int32) []Chunk { panic("") }
+func Split(data []byte, chunkSize uint64) []Chunk { panic("") }
 
 // splits `data` in default-sized chunks and calculates Merkle root out of them
 func MerkleHash(data []byte) Digest { panic("") }
 
 // splits `data` in default-sized chunks and calculates proof for the specified range
-func CreateMerkleProof(data []byte, offset int32, length int32) MerkleProof { panic("") }
+func CreateMerkleProof(data []byte, offset uint64, length uint64) MerkleProof { panic("") }
 
 // checks merkle proof for the range of default-sized chunks
 func VerifyMerkleProof(data []byte, proof MerkleProof, merkleRoot Digest) bool { panic("") }
